+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/agent/ValidatorVpTest.kt
@@ -36,6 +36,7 @@ class ValidatorVpTest : FreeSpec({
 
     lateinit var validator: Validator
     lateinit var issuer: Issuer
+    lateinit var statusListIssuer: StatusListIssuer
     lateinit var issuerCredentialStore: IssuerCredentialStore
     lateinit var holder: HolderAgent
     lateinit var holderCredentialStore: SubjectCredentialStore
@@ -49,11 +50,11 @@ class ValidatorVpTest : FreeSpec({
         validator = Validator(
             resolveStatusListToken = {
                 if (Random.nextBoolean()) StatusListToken.StatusListJwt(
-                    issuer.issueStatusListJwt(),
+                    statusListIssuer.issueStatusListJwt(),
                     resolvedAt = Clock.System.now()
                 ) else {
                     StatusListToken.StatusListCwt(
-                        issuer.issueStatusListCwt(),
+                        statusListIssuer.issueStatusListCwt(),
                         resolvedAt = Clock.System.now()
                     )
                 }
@@ -64,7 +65,7 @@ class ValidatorVpTest : FreeSpec({
             EphemeralKeyWithoutCert(),
             validator = validator,
             issuerCredentialStore = issuerCredentialStore,
-        )
+        ).also { statusListIssuer = it }
         holderCredentialStore = InMemorySubjectCredentialStore()
         holderKeyMaterial = EphemeralKeyWithoutCert()
         holder = HolderAgent(