+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 Release 5.8.0:
- - tbd
+ - Code organization:
+   - Extract interface `StatusListIssuer` out of `Issuer` to separate credential issuing and status list management
 
 Release 5.7.1:
  - Signum 3.16.3/Supreme 0.8.3 to fix certificate encoding in JWS header